A couple more python 3 nuances.
The `universal_newlines` flag is ignored by python 2, but without it, python 3
yields a bytestring here, which is a monkey wrench in whatever uses the result.

The call to "zip" in python 3 yields an iterator, which can only be consumed
once, and isn't memoized, so the subsequent uses of `cls` fail to do anything.